Update __init__.py to set default value for logoName attribute
This addresses issue #440 in the alarmdotcom project. pyalarmdotcom/alarmdotcom#440
Issue #145 here.

Something changed in the response alarm.com returns on login, leaving out the "logoName" attribute. This causes pyalarmdotcomajax to terminate the authentication process as failed.

This change is to fill in a default value for that attribute in the case that it is missing from the alarm.com response.